Add result and option transformers
